feat: 가게 즐겨찾기 등록 추가(삭제는 오류 발생)

+++ src/main/java/com/fivemybab/ittabab/store/command/application/controller/StoreMenuController.java
@@ -1,5 +1,7 @@
 package com.fivemybab.ittabab.store.command.application.controller;
 
+import com.fivemybab.ittabab.exception.NotFoundException;
+import com.fivemybab.ittabab.security.util.CustomUserDetails;
 import com.fivemybab.ittabab.store.command.application.dto.CreateStoreMenuDto;
 import com.fivemybab.ittabab.store.command.application.dto.UpdateStoreMenuDto;
 import com.fivemybab.ittabab.store.command.application.service.StoreMenuService;
@@ -8,6 +10,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -22,15 +25,29 @@ public class StoreMenuController {
     /* 가게 메뉴 등록하기 */
     @Operation(summary = "메뉴 등록")
     @PostMapping
-    public ResponseEntity<CreateStoreMenuDto> CreateStoreMenu(@RequestBody CreateStoreMenuDto createStoreMenuDto) {
+    public ResponseEntity<String> CreateStoreMenu(@RequestBody CreateStoreMenuDto createStoreMenuDto, @AuthenticationPrincipal CustomUserDetails loginUser) {
+
+        // 로그인되지 않았거나, userId가 null인 경우 예외 처리
+        if (loginUser == null || loginUser.getUserId() == null) {
+            throw new NotFoundException("로그인이 필요합니다.");
+        }
+
         storeMenuService.createStoreMenu(createStoreMenuDto);
-        return new ResponseEntity<>(createStoreMenuDto, HttpStatus.CREATED);
+        return new ResponseEntity<>("메뉴 등록 완료", HttpStatus.CREATED);
     }
 
     /* 가게 메뉴 수정하기 */
     @Operation(summary = "메뉴 수정")
     @PutMapping("/{menuId}")
-    public ResponseEntity<UpdateStoreMenuDto> UpdateStoreMenu(@PathVariable Long menuId, @RequestBody UpdateStoreMenuDto updateStoreMenuDto) {
+    public ResponseEntity<UpdateStoreMenuDto> UpdateStoreMenu(@PathVariable Long menuId, @RequestBody UpdateStoreMenuDto updateStoreMenuDto,
+    @AuthenticationPrincipal CustomUserDetails loginUser)
+    {
+
+        // 로그인되지 않았거나, userId가 null인 경우 예외 처리
+        if (loginUser == null || loginUser.getUserId() == null) {
+            throw new NotFoundException("로그인이 필요합니다.");
+        }
+
         storeMenuService.updateStoreMenu(menuId, updateStoreMenuDto);
         return new ResponseEntity<>(updateStoreMenuDto, HttpStatus.OK);
     }
@@ -39,7 +56,13 @@ public class StoreMenuController {
     /* 가게 메뉴 삭제하기 */
     @Operation(summary = "메뉴 삭제")
     @DeleteMapping("/{menuId}")
-    public ResponseEntity<Void> DeleteStoreMenu(@PathVariable Long menuId) {
+    public ResponseEntity<Void> DeleteStoreMenu(@PathVariable Long menuId, @AuthenticationPrincipal CustomUserDetails loginUser) {
+
+        // 로그인되지 않았거나, userId가 null인 경우 예외 처리
+        if (loginUser == null || loginUser.getUserId() == null) {
+            throw new NotFoundException("로그인이 필요합니다.");
+        }
+
         storeMenuService.deleteStoreMenu(menuId);
         return ResponseEntity.noContent().build();
     }
